+++ packages/eslint-config-react-typescript/index.js
@@ -1,3 +1,22 @@
 module.exports = {
   extends: ['@winwangqi/eslint-config-react', '@winwangqi/eslint-config-typescript'].map(require.resolve),
+  rules: {
+    "import/extensions": [
+      "error",
+      "ignorePackages",
+      {
+        "js": "never",
+        "jsx": "never",
+        "ts": "never",
+        "tsx": "never"
+      }
+    ],
+  },
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
+      },
+    },
+  },
 }