+++ packages/eslint-config-react-typescript/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @winwangqi/eslint-config-react-typescript
 
+## 1.1.2
+
+### Patch Changes
+
+- add import extensions
+- Updated dependencies
+  - @winwangqi/eslint-config-react@1.1.2
+  - @winwangqi/eslint-config-typescript@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes