+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @winwangqi/eslint-config-react
 
+## 1.1.2
+
+### Patch Changes
+
+- add import extensions
+- Updated dependencies
+  - @winwangqi/eslint-config-core@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes