+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @winwangqi/eslint-config
 
+## 1.1.2
+
+### Patch Changes
+
+- add import extensions
+- Updated dependencies
+  - @winwangqi/eslint-config-react-typescript@1.1.2
+
 ## 1.1.1
 
 ### Patch Changes